+++ widgets/events/components/events-list/events-list.module.css
@@ -1,13 +1,10 @@
-.statusText {
-	color: var(--wpds-color-fg-content-neutral-weak);
-	margin: 0;
+.root {
+	flex: 1;
+	min-width: 0;
+	min-height: 0;
 }
 
-.listItem {
-	inline-size: 100%;
-}
-
-.listItemIcon {
+.eventIcon {
 	inline-size: calc(var(--wpds-dimension-base) * 9);
 	block-size: calc(var(--wpds-dimension-base) * 9);
 	display: inline-flex;
@@ -16,10 +13,15 @@
 	border-radius: var(--wpds-border-radius-lg);
 	background: var(--wpds-color-bg-track-neutral-weak);
 	color: var(--wpds-color-fg-content-neutral-weak);
-	flex-shrink: 0;
 }
 
-.listItemContent {
-	flex: 1;
-	min-inline-size: 0;
+.titleLink {
+	min-width: 0;
+	overflow: hidden;
+	white-space: nowrap;
+	text-overflow: ellipsis;
+}
+
+.description {
+	color: var(--wpds-color-fg-content-neutral-weak);
 }
