+++ widgets/events/components/events-list/index.ts
