--- widgets/events/components/list/index.ts
@@ -1,1 +1,0 @@
-export { List, type ListItem } from './list';
